requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=-
#
aiofiles==24.1.0
    # via asyncpraw
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via
    #   asyncpraw
    #   asyncprawcore
aiosignal==1.3.2
    # via aiohttp
aiosqlite==0.17.0
    # via asyncpraw
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via starlette
asyncpraw==7.8.1
    # via -r requirements.in
asyncprawcore==2.4.0
    # via asyncpraw
attrs==25.1.0
    # via aiohttp
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
fastapi==0.115.7
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
    #   yarl
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
numpy==2.2.2
    # via
    #   -r requirements.in
    #   pandas
pandas==2.2.3
    # via -r requirements.in
pexpect==4.9.0
    # via -r requirements.in
praw==7.8.1
    # via -r requirements.in
prawcore==2.4.0
    # via praw
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
ptyprocess==0.7.0
    # via pexpect
pydantic==2.10.6
    # via fastapi
pydantic-core==2.27.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via pandas
requests==2.32.3
    # via
    #   -r requirements.in
    #   prawcore
    #   update-checker
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.45.3
    # via fastapi
typing-extensions==4.12.2
    # via
    #   aiosqlite
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
tzdata==2025.1
    # via pandas
update-checker==0.18.0
    # via
    #   asyncpraw
    #   praw
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via -r requirements.in
websocket-client==1.8.0
    # via praw
yarl==1.18.3
    # via
    #   aiohttp
    #   asyncprawcore
